+++ gherkin/java/src/test/java/io/cucumber/gherkin/ParserTest.java
@@ -11,10 +11,10 @@ public class ParserTest {
     @Test
     public void change_default_language() {
         TokenMatcher matcher = new TokenMatcher("no");
-        IdGenerator idGenerator = new IdGenerator.Incrementing();
-        Parser<GherkinDocument> parser = new Parser<>(new GherkinDocumentBuilder(idGenerator));
+        IdGenerator idGenerator = new IncrementingIdGenerator();
+        Parser<GherkinDocument> parser = new Parser<>(new GherkinDocumentBuilder(idGenerator, "test.feature"));
 
-        GherkinDocument gherkinDocument = parser.parse("Egenskap: i18n support\n", matcher);
-        assertEquals("no", gherkinDocument.getFeature().getLanguage());
+        GherkinDocument gherkinDocument = parser.parse("Egenskap: i18n support\n", matcher, "test.feature");
+        assertEquals("no", gherkinDocument.getFeature().get().getLanguage());
     }
 }