+++ react/javascript/src/components/customise/index.ts
@@ -1,2 +1,0 @@
-export * from './customRendering'
-export * from './theming'